Add login settings list
