+++ src/scenes/Login/LoginSettings/index.js
@@ -1,0 +1,3 @@
+import LoginSettings from './LoginSettings'
+
+export default LoginSettings
