+++ src/scenes/Login/LoginSettings/strings/en.js
@@ -1,0 +1,5 @@
+export default {
+  'network': 'Network',
+  'language': 'Language',
+  'faq': 'FAQ'
+}
