+++ src/scenes/Login/LoginSettings/strings/index.js
@@ -1,0 +1,5 @@
+import LocalizedStrings from 'react-native-localization'
+import en from './en'
+import ru from './ru'
+
+export default new LocalizedStrings({ en, ru })
