+++ src/scenes/Login/LoginSettings/strings/ru.js
@@ -1,0 +1,5 @@
+export default {
+  'network': 'Сеть',
+  'language': 'Язык',
+  'faq': 'Частые вопросы'
+}
